+++ packages/openai/src/chat/openai-chat-prepare-tools.ts
@@ -3,9 +3,10 @@ import {
   LanguageModelV2CallWarning,
   UnsupportedFunctionalityError,
 } from '@ai-sdk/provider';
-import { fileSearchArgsSchema } from '../tool/file-search';
-import { webSearchPreviewArgsSchema } from '../tool/web-search-preview';
-import { OpenAIChatToolChoice, OpenAIChatTools } from './openai-chat-types';
+import {
+  OpenAIChatToolChoice,
+  OpenAIChatFunctionTool,
+} from './openai-chat-types';
 
 export function prepareChatTools({
   tools,
@@ -18,7 +19,7 @@ export function prepareChatTools({
   structuredOutputs: boolean;
   strictJsonSchema: boolean;
 }): {
-  tools?: OpenAIChatTools;
+  tools?: OpenAIChatFunctionTool[];
   toolChoice?: OpenAIChatToolChoice;
   toolWarnings: Array<LanguageModelV2CallWarning>;
 } {
@@ -31,7 +32,7 @@ export function prepareChatTools({
     return { tools: undefined, toolChoice: undefined, toolWarnings };
   }
 
-  const openaiTools: OpenAIChatTools = [];
+  const openaiTools: OpenAIChatFunctionTool[] = [];
 
   for (const tool of tools) {
     switch (tool.type) {
@@ -45,35 +46,6 @@ export function prepareChatTools({
             strict: structuredOutputs ? strictJsonSchema : undefined,
           },
         });
-        break;
-      case 'provider-defined':
-        switch (tool.id) {
-          case 'openai.file_search': {
-            const args = fileSearchArgsSchema.parse(tool.args);
-            openaiTools.push({
-              type: 'file_search',
-              vector_store_ids: args.vectorStoreIds,
-              max_num_results: args.maxNumResults,
-              ranking_options: args.ranking
-                ? { ranker: args.ranking.ranker }
-                : undefined,
-              filters: args.filters,
-            });
-            break;
-          }
-          case 'openai.web_search_preview': {
-            const args = webSearchPreviewArgsSchema.parse(tool.args);
-            openaiTools.push({
-              type: 'web_search_preview',
-              search_context_size: args.searchContextSize,
-              user_location: args.userLocation,
-            });
-            break;
-          }
-          default:
-            toolWarnings.push({ type: 'unsupported-tool', tool });
-            break;
-        }
         break;
       default:
         toolWarnings.push({ type: 'unsupported-tool', tool });