+++ packages/openai/src/transcription/openai-transcription-model.test.ts
@@ -1,9 +1,13 @@
-import { createTestServer } from '@ai-sdk/provider-utils/test';
+import { createTestServer } from '@ai-sdk/test-server/with-vitest';
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
 import { createOpenAI } from '../openai-provider';
 import { OpenAITranscriptionModel } from './openai-transcription-model';
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect, vi } from 'vitest';
+
+vi.mock('../version', () => ({
+  VERSION: '0.0.0-test',
+}));
 
 const audioData = await readFile(
   path.join(__dirname, 'transcription-test.mp3'),
@@ -114,6 +118,10 @@ describe('doGenerate', () => {
       'openai-organization': 'test-organization',
       'openai-project': 'test-project',
     });
+
+    expect(server.calls[0].requestUserAgent).toContain(
+      `ai-sdk/openai/0.0.0-test`,
+    );
   });
 
   it('should extract the transcription text', async () => {
@@ -206,7 +214,7 @@ describe('doGenerate', () => {
         "model": "whisper-1",
         "response_format": "verbose_json",
         "temperature": "0",
-        "timestamp_granularities": "word",
+        "timestamp_granularities[]": "word",
       }
     `);
   });
@@ -235,7 +243,7 @@ describe('doGenerate', () => {
         "model": "gpt-4o-transcribe",
         "response_format": "json",
         "temperature": "0",
-        "timestamp_granularities": "word",
+        "timestamp_granularities[]": "word",
       }
     `);
   });
@@ -263,7 +271,7 @@ describe('doGenerate', () => {
         "model": "whisper-1",
         "response_format": "verbose_json",
         "temperature": "0",
-        "timestamp_granularities": "segment",
+        "timestamp_granularities[]": "segment",
       }
     `);
   });